add delete picture from blog and delete blog options

+++ src/main/java/com/exam/blog/service/BlogService.java
@@ -2,6 +2,7 @@ package com.exam.blog.service;
 
 
 import com.exam.blog.models.Blog;
+import com.exam.blog.models.Comment;
 import com.exam.blog.models.Picture;
 import com.exam.blog.models.User;
 import com.exam.blog.repository.BlogRepo;
@@ -12,9 +13,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.IOException;
 import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -27,6 +26,12 @@ public class BlogService {
     private BlogRepo blogRepo;
     private UserRepoImpl userRepo;
     private PictureService pictureService;
+    private CommentService commentService;
+
+    @Autowired
+    public void setCommentService(CommentService commentService) {
+        this.commentService = commentService;
+    }
 
     @Autowired
     public void setPictureService(PictureService pictureService) {
@@ -87,9 +92,9 @@ public class BlogService {
 
             picture.setId(null);
 
-           addProperties(blog, idUser);
+            addProperties(blog, idUser);
 
-            pictureService.uploadPictureImage(idUser, blog, image, picture);
+            pictureService.loadPictureImage(idUser, blog, image, picture);
 
             picture.setBlog(blog);
 
@@ -120,7 +125,7 @@ public class BlogService {
 
     public void updatePropertiesBlog(Blog blog, Picture picture, Long idUser, MultipartFile image) throws IOException{
         if(picture != null){
-           pictureService.uploadPictureImage(idUser, blog, image, picture);
+           pictureService.updateBlogLoadPictureImage(idUser, blog, image, picture);
 
         }
     }
@@ -130,40 +135,65 @@ public class BlogService {
         Blog blogDB = getById(blog.getId());
         if(blogDB != null) {
 
-            blog.setRating(blogDB.getRating());
-            blog.setDate_create_blog(blogDB.getDate_create_blog());
-            blog.setUser(blogDB.getUser());
-            blog.setBan_blog(blogDB.getBan_blog());
-            blog.setComments(blogDB.getComments());
-            blog.setPictures(blogDB.getPictures());
+            blogDB.setAnnouncement(blog.getAnnouncement());
+            blogDB.setTitle(blog.getTitle());
+            blogDB.setFull_text(blog.getFull_text());
 
-            if(picture.getName() != null){
-
-                blog.getPictures().toArray(Picture[]::new)[0].setName(picture.getName());
-            }
             if(image != null){
-                File folder = new File(uploadPicturePath + "/" + idUser + "/" + blog.getId());
+                File folder = new File(uploadPicturePath + "/" + idUser + "/" + blogDB.getId());
 
                 File[] files = folder.listFiles();
 
-                if(files.length != 0){
-                    for(File file : files){
-                        file.delete();
+                if(files != null) {
+                    if (files.length != 0) {
+                        for (File file : files) {
+                            file.delete();
+                        }
                     }
-                    image.transferTo(new File(folder, Objects.requireNonNull(image.getOriginalFilename())));
                 }
             }
 
-//            updatePropertiesBlog(blog, picture, idUser, image);
+            updatePropertiesBlog(blogDB, picture, idUser, image);
 
-            update(blog);
-            return "redirect:/user/blog/" + idUser + "/" + blog.getId() + "/" + blog.getUser().getId();
+            Picture pictureDB = pictureService.getById(picture.getId());
+
+            blogDB.getPictures().remove(0);
+
+            blogDB.getPictures().add(pictureDB);
+
+            update(blogDB);
+
+            return "redirect:/user/blog/" + idUser + "/" + blogDB.getId() + "/" + blogDB.getUser().getId();
 
         } else {
 
-            return "redirect:/user/update-blog/" +idUser + "/" + blog.getId();
+            return "redirect:/user/update-blog/" +idUser + "/" + blogDB.getId();
 
         }
     }
 
+    public void deleteBlog(User userDB, Blog blogDB){
+
+        for(Blog blog : userDB.getBlogs()) {
+            if (blog.getId() == blogDB.getId()) {
+                userDB.getBlogs().remove(blog);
+            }
+        }
+
+        for (Comment comment: commentService.findAllCommentsBlog()) {
+            for (Comment commentBlog : blogDB.getComments()) {
+                if (comment.getId() == commentBlog.getId()) {
+                    commentService.delete(comment.getId());
+                }
+            }
+        }
+
+        for(Picture picture : pictureService.findAllPicture()){
+            if(picture.getId() == blogDB.getPictures().get(0).getId()){
+                blogDB.getPictures().remove(picture);
+            }
+        }
+
+        delete(blogDB.getId());
+    }
 }
